requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
disnake==2.10.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
pillow==11.1.0
playwright==1.50.0
pluggy==1.5.0
propcache==0.2.1
pycparser==2.22
pyee==12.1.1
PyNaCl==1.5.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-base-url==2.1.0
pytest-mock==3.14.0
pytest-playwright==0.7.0
python-dotenv==1.0.1
python-slugify==8.0.4
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
spotipy==2.25.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
youtube-search-python==1.6.6
yt-dlp==2025.1.26
